reasonable-productivity-backend/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
drf-nested-routers==0.92.5
flake8==3.8.4
importlib-metadata==3.3.0
mccabe==0.6.1
Pillow==8.1.0
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.5
sqlparse==0.4.1
zipp==3.4.0
